feat(login) : readDB
