+++ app.py
@@ -1,26 +1,69 @@
 import chatbot
-from flask import Flask, render_template, request, redirect
+from flask import Flask, render_template, request, redirect, session, url_for, jsonify, flash
 from jinja2 import Environment, FileSystemLoader, Template
 from pymongo import MongoClient
+from decouple import config
 
 app = Flask(__name__)
 
-client = MongoClient('mongodb://localhost:27017/')
+# 시크릿키 설정
+app.secret_key = config('APP_SECRET_KEY')
+
+client = MongoClient('localhost', 27017)
 db = client['WEEK00_TEAM7']
-collection = db['A']
+collection = db['users']
 collection2 = db['B']
 collection3 = db['C']
 
 @app.route('/')
-def loginRender():
+def entrypoint():
+    if 'username' in session:
+        return redirect('/main')
+    else:
+        return redirect('/login')
+
+@app.route('/login', methods=['GET', 'POST'])
+def login():
+    if request.method == "GET":
+        return render_template("login.html")
+    # POST 일 때
+    userId_receive = request.form.get('username')
+    userPwd_receive = request.form.get('password')
+    if userId_receive == "":
+        flash("아이디를 입력해주세요.")
+    if userPwd_receive == "":
+        flash("비밀번호를 입력해주세요.")
+        return render_template("login.html")
+    
+    user = db.users.find_one({'id':userId_receive}, {'pwd':userPwd_receive})
+    if user is None:
+        flash("존재하지 않는 유저입니다.")
+        return render_template("login.html")
+    session["username"] = userId_receive
     return render_template("login.html")
 
-@app.route('/register', methods=['GET', 'POST'])
+@app.route('/logout', methods=['GET'])
+def logout():
+    session.pop("username")
+    return redirect('/login')
+
+@app.route('/register', methods=["GET", "POST"])
 def registerRender():
-    #GET
     if request.method == "POST":
-        return render_template("login.html")
-    else :
+        userId_receive = request.form.get('username')
+        userPwd_receive = request.form.get('password')
+        
+        user = {
+            'id': userId_receive,
+            'pwd': userPwd_receive
+        }
+        if userId_receive == '' or userPwd_receive == '':
+            db.users.insert_one()
+        else:
+            db.users.insert_one(user)
+            return jsonify({'result':'success', 'msg':'POST 연결되었습니다!'})
+        print(user)
+    else:
         return render_template("register.html")
 
 @app.route('/main', methods=["GET", "POST"])
@@ -45,13 +88,7 @@ def mainRender():
     # POST
     json = request.get_json()
     completion = chatbot.create_completion(json)
-
     return {"message": completion.choices[0].message.content}
-
-
-@app.route('/regist' , methods=['POST'])
-def login():
-    return 'hello'
 
 if __name__ == '__main__':
     app.run('0.0.0.0', port=4000, debug=True)
