+++ templates/login.html
@@ -9,7 +9,17 @@
 <body>
     <div>This is login webpage</div>
 
-    <form action="/main">
+    {% with messages = get_flashed_messages() %}
+        {% if messages %}
+            <ul class=flashes>
+               {% for message in messages %}
+                    <li>{{ message }}</li>
+                {% endfor %}
+            </ul>
+        {% endif %}
+    {% endwith %}
+
+    <form action="/login" method="POST">
 
         <div class="input-box">
             <input id="username" type="text" name="username" placeholder="아이디를 입력하세요.">
@@ -27,7 +37,6 @@
             <button type="submit">회원가입</button>
         </form>
     </div>
-
 
 </body>
 </html>
